refactor check suite forwarding for better readability
